Séparaation du programme SPI setup
